requirements.txt
streamlit
torch
transformers
Pillow
plotly
pandas
numpy
accelerate